backend/requirements.txt
fastapi
uvicorn
pillow
opencv-python-headless
opencv-python
shutil